Add default rustfmt configuration and remove vim function
rustfmt defaults to the 2015 edition, which is quite different from 2018
onwards.  This provides a reasonable default as a fallback that is
largely consistent with modern projects.  This also provides a consitent
experience outside of vim.

This doesn't create the necessary symlink at ~/.config/rustfmt because I
want to get rid of the rust executable managing those.  This must be
done manually.

Taken from:
https://github.com/rust-lang/rustfmt/blob/master/Configurations.md